lab_supabase/pyproject.toml
[project]
name = "lab-supabase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.24.0",
]

